Refactor(Bank Reconciliation): segregate js
